Fixed some typos. Changed import order in test_fonts.py. Reverted version to 1.0.1. Shorted the guess_font_filename code Moved font related code to separate file.
